feat(dashboard): merge web chat into dashboard app
Consolidate @corvus/chat into @corvus/dashboard, making dashboard the
sole operator-facing Vue/Vite app. Chat becomes a first-class tab
alongside Config, Sessions, and Memory views.

- Create ChatWorkspace component from chat App.vue
- Add useChatGateway adapter bridging useConfig auth to ChatGateway interface
- Port useChat composable with ChatGateway interface instead of useGateway
- Move chat components (ChatMessage, ToolApprovalCard, HealthIndicator, SessionSidebar)
- Add Chat tab to dashboard navigation
- Port all chat tests (47 files, 306 tests passing)
- Remove @corvus/chat package and all chat-specific scripts/targets
- Update Makefile, build.gradle.kts, package.json, README, AGENTS.md
