+++ .agents/AGENTS.md
@@ -123,14 +123,11 @@ If using Gradle directly for Rust tasks, include `-PenableRustTasks=true`.
 
 - `make web-install`, `make web-build-all`, `make web-test-all`, `make web-check-all`
 - `make docs-dev` / `make docs-build` / `make docs-check`
-- `make chat-dev` / `make chat-build` / `make chat-check` / `make chat-test`
 - `make dashboard-dev` / `make dashboard-build` / `make dashboard-check` / `make dashboard-test`
 - `make marketing-dev` / `make marketing-build` / `make marketing-check`
   Direct pnpm examples from `clients/web`:
-- `pnpm test`, `pnpm test:chat`, `pnpm test:dashboard`, `pnpm test:dashboard:e2e`
+- `pnpm test`, `pnpm test:dashboard`, `pnpm test:dashboard:e2e`
   Single web test examples:
-- `pnpm --dir clients/web --filter @corvus/chat test -- src/path/to/file.test.ts`
-- `pnpm --dir clients/web --filter @corvus/chat test -- src/path/to/file.test.ts -t "case name"`
 - `pnpm --dir clients/web --filter @corvus/dashboard test -- src/path/to/file.test.ts`
 - `pnpm --dir clients/web --filter @corvus/dashboard exec playwright test e2e/file.spec.ts --grep "case name"`
