--- clients/web/apps/chat/components.json
@@ -1,15 +1,0 @@
-{
-  "$schema": "https://www.shadcn-vue.com/schema.json",
-  "style": "default",
-  "typescript": true,
-  "tailwind": {
-    "config": "",
-    "css": "src/style.css",
-    "baseColor": "slate",
-    "cssVariables": true
-  },
-  "aliases": {
-    "components": "@/components",
-    "utils": "@/lib/utils"
-  }
-}
