--- clients/web/apps/chat/package.json
@@ -1,50 +1,0 @@
-{
-  "name": "@corvus/chat",
-  "version": "1.0.0",
-  "private": true,
-  "description": "Corvus Chat — conversational AI interface",
-  "type": "module",
-  "engines": {
-    "node": ">=22.0.0"
-  },
-  "scripts": {
-    "dev": "portless run --name chat vite --config vite.config.ts",
-    "build": "vue-tsc -b && vite build --config vite.config.ts",
-    "preview": "portless run --name chat vite preview --config vite.config.ts",
-    "format": "biome format --write src package.json components.json tsconfig*.json vite.config.ts index.html",
-    "check": "biome check src package.json components.json tsconfig*.json vite.config.ts index.html",
-    "test": "vitest --config vite.config.ts --run",
-    "test:coverage": "vitest --config vite.config.ts --run --coverage"
-  },
-  "dependencies": {
-    "@corvus/locales": "workspace:*",
-    "@corvus/shared": "workspace:*",
-    "@corvus/ui": "workspace:*",
-    "@fontsource-variable/space-grotesk": "catalog:",
-    "@fontsource/space-mono": "catalog:",
-    "@fontsource/doto": "catalog:",
-    "clsx": "^2.1.1",
-    "tailwind-merge": "catalog:",
-    "vue": "catalog:",
-    "vue-i18n": "catalog:"
-  },
-  "devDependencies": {
-    "@biomejs/biome": "catalog:",
-    "@tailwindcss/postcss": "catalog:",
-    "@tsconfig/node22": "catalog:",
-    "@types/node": "catalog:",
-    "@vitejs/plugin-vue": "catalog:",
-    "@vitest/coverage-v8": "catalog:",
-    "@vue/compiler-dom": "catalog:",
-    "@vue/test-utils": "catalog:",
-    "@vue/tsconfig": "catalog:",
-    "happy-dom": "catalog:",
-    "postcss": "catalog:",
-    "portless": "catalog:",
-    "tailwindcss": "catalog:",
-    "typescript": "catalog:",
-    "vite": "catalog:",
-    "vitest": "catalog:",
-    "vue-tsc": "catalog:"
-  }
-}
