--- clients/web/apps/chat/src/style.css
@@ -1,37 +1,0 @@
-@import url("@corvus/shared/nothing-shell.css");
-@import url("tailwindcss");
-@import url("@corvus/shared/tailwind-bridge.css");
-
-/* ── Animations ─────────────────────────────────────────────── */
-
-@keyframes fade-in {
-  from {
-    opacity: 0;
-    transform: translateY(8px);
-  }
-
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-@keyframes slide-up {
-  from {
-    opacity: 0;
-    transform: translateY(16px);
-  }
-
-  to {
-    opacity: 1;
-    transform: translateY(0);
-  }
-}
-
-.animate-fade-in {
-  animation: fade-in 0.3s ease-out forwards;
-}
-
-.animate-slide-up {
-  animation: slide-up 0.4s cubic-bezier(0.2, 0.8, 0.2, 1) forwards;
-}
