--- clients/web/apps/chat/src/types/chat.ts
@@ -1,35 +1,0 @@
-export interface Message {
-  id: number;
-  role: "user" | "assistant";
-  content: string;
-  status?: "streaming" | "complete" | "error";
-  recalledMemoryKeys?: string[];
-}
-
-export interface StreamChunkEvent {
-  text: string;
-}
-
-export interface StreamDoneEvent {
-  message_id: string;
-  session_id?: string;
-  recalled_memory_keys?: string[];
-}
-
-export interface StreamErrorEvent {
-  code: string;
-  message: string;
-}
-
-export interface SessionListItem {
-  id: string;
-  started_at: string;
-  ended_at: string | null;
-  message_count: number;
-  last_activity: string;
-}
-
-export interface SessionListResponse {
-  sessions: SessionListItem[];
-  total: number;
-}
