--- clients/web/apps/chat/tsconfig.json
@@ -1,14 +1,0 @@
-{
-  "files": [],
-  "references": [
-    {
-      "path": "./tsconfig.app.json"
-    },
-    {
-      "path": "./tsconfig.node.json"
-    },
-    {
-      "path": "../../packages/locales/tsconfig.json"
-    }
-  ]
-}
