--- clients/web/apps/chat/tsconfig.node.json
@@ -1,11 +1,0 @@
-{
-  "extends": "@tsconfig/node22/tsconfig.json",
-  "compilerOptions": {
-    "tsBuildInfoFile": "./node_modules/.tmp/tsconfig.node.tsbuildinfo",
-    "module": "ESNext",
-    "moduleResolution": "Bundler",
-    "types": ["node"],
-    "lib": ["ESNext", "ESNext.Disposable"]
-  },
-  "include": ["vite.config.ts"]
-}
