+++ clients/web/apps/dashboard/src/components/chat/ChatMessage.vue
