+++ clients/web/apps/dashboard/src/components/chat/HealthIndicator.vue
