+++ clients/web/apps/dashboard/src/components/chat/SessionSidebar.vue
