+++ clients/web/apps/dashboard/src/components/chat/ToolApprovalCard.spec.ts
