+++ clients/web/apps/dashboard/src/components/chat/ToolApprovalCard.vue
