+++ clients/web/build.gradle.kts
@@ -72,7 +72,6 @@ data class WebAppConfig(
 val appConfigs = mapOf(
   "docs" to WebAppConfig("docs", "dist", 4321),
   "marketing" to WebAppConfig("marketing", "dist", 9988),
-  "chat" to WebAppConfig("chat", "dist", 4323),
   "dashboard" to WebAppConfig("dashboard", "dist", 4324),
 )
