+++ clients/web/package.json
@@ -2,7 +2,7 @@
   "name": "@corvus/web-monorepo",
   "version": "1.0.0",
   "private": true,
-  "description": "Corvus Web Apps Monorepo - Marketing, Docs, Chat, Dashboard",
+  "description": "Corvus Web Apps Monorepo - Marketing, Docs, Dashboard",
   "type": "module",
   "scripts": {
     "update-deps": "pnpm update -i -r --latest",
@@ -10,13 +10,11 @@
     "build:docs": "pnpm --filter @corvus/docs run build",
     "build:marketing": "pnpm --filter @corvus/marketing run build",
     "build:landing": "pnpm run build:marketing",
-    "build:chat": "pnpm --filter @corvus/chat run build",
     "build:dashboard": "pnpm --filter @corvus/dashboard run build",
     "dev": "pnpm --parallel -r run dev",
     "dev:docs": "pnpm --filter @corvus/docs run dev",
     "dev:marketing": "pnpm --filter @corvus/marketing run dev",
     "dev:landing": "pnpm run dev:marketing",
-    "dev:chat": "pnpm --filter @corvus/chat run dev",
     "dev:dashboard": "pnpm --filter @corvus/dashboard run dev",
     "format": "pnpm -r run format",
     "format:styles": "stylelint \"apps/*/src/**/*.css\" \"packages/**/*.css\" --fix",
@@ -24,7 +22,6 @@
     "check:styles": "stylelint \"apps/*/src/**/*.css\" \"packages/**/*.css\"",
     "clean": "pnpm -r run clean",
     "test": "pnpm -r run test",
-    "test:chat": "pnpm --filter @corvus/chat run test",
     "test:dashboard": "pnpm --filter @corvus/dashboard run test",
     "test:dashboard:e2e": "pnpm --filter @corvus/dashboard run test:e2e",
     "test:dashboard:e2e:install": "pnpm --filter @corvus/dashboard run test:e2e:install"
